feat(ext): extend fetch base
